# packages/trip-check-client/openapi.current.json
{
  "components": {
    "schemas": {
...
      "ConsentUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "ConsentUpdateRequest",
        "type": "object"
      },
...
      "DataConsentView": {
        "additionalProperties": false,
        "properties": {
          "feedback_enabled": {
            "default": false,
            "title": "Feedback Enabled",
            "type": "boolean"
          },
          "memory_enabled": {
            "default": false,
            "title": "Memory Enabled",
            "type": "boolean"
          },
          "training_eval_enabled": {
            "default": false,
            "title": "Training Eval Enabled",
            "type": "boolean"
          }
        },
        "title": "DataConsentView",
        "type": "object"
      },
...
      "FeedbackAcceptedView": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "const": "RECORDED",
            "default": "RECORDED",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "FeedbackAcceptedView",
        "type": "object"
      },
      "FeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "event_type": {
            "enum": [
              "CORRECTION",
              "ADOPTED",
              "REJECTED",
              "VOLUNTARY"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "subject_ref": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Ref"
          },
          "subject_type": {
            "enum": [
              "TRIP",
              "ACTIVITY",
              "KNOWLEDGE_SUGGESTION"
            ],
            "title": "Subject Type",
            "type": "string"
          }
        },
        "required": [
          "event_type",
          "subject_type"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PreferenceMemoryView": {
        "additionalProperties": false,
        "properties": {
          "dining_preferences": {
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "title": "Dining Preferences",
            "type": "array"
          },
          "hotel_preferences": {
            "items": {
              "type": "string"
            },
            "maxItems": 3,
            "title": "Hotel Preferences",
            "type": "array"
          },
          "intensity": {
            "anyOf": [
              {
                "enum": [
                  "RELAXED",
                  "BALANCED",
                  "FULL"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intensity"
          },
          "preferred_start_time": {
            "anyOf": [
              {
                "pattern": "^([01][0-9]|2[0-3]):[0-5][0-9]$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Start Time"
          },
          "walking_tolerance_minutes": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 5.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Walking Tolerance Minutes"
          }
        },
        "title": "PreferenceMemoryView",
        "type": "object"
      },
...
      "ShareCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "expires_in_days": {
            "default": 7,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Expires In Days",
            "type": "integer"
          }
        },
        "title": "ShareCreateRequest",
        "type": "object"
      },
      "ShareCreatedView": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "share_url": {
            "title": "Share Url",
            "type": "string"
          }
        },
        "required": [
          "share_url",
          "expires_at"
        ],
        "title": "ShareCreatedView",
        "type": "object"
      },
      "ShareExchangeRequest": {
        "additionalProperties": false,
        "properties": {
          "secret": {
            "maxLength": 100,
            "minLength": 40,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "secret"
        ],
        "title": "ShareExchangeRequest",
        "type": "object"
      },
      "ShareListItemView": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "share_ref": {
            "title": "Share Ref",
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "REVOKED",
              "EXPIRED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "share_ref",
          "expires_at",
          "status"
        ],
        "title": "ShareListItemView",
        "type": "object"
      },
      "ShareProjectionView": {
        "additionalProperties": false,
        "properties": {
          "accommodation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodation"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/SharedDayView"
            },
            "title": "Days",
            "type": "array"
          },
          "destination": {
            "title": "Destination",
            "type": "string"
          },
          "message": {
            "default": "这是朋友分享的只读行程。",
            "title": "Message",
            "type": "string"
          },
          "party_size": {
            "title": "Party Size",
            "type": "string"
          },
          "schedule": {
            "title": "Schedule",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "destination",
          "schedule",
          "party_size",
          "days"
        ],
        "title": "ShareProjectionView",
        "type": "object"
      },
      "SharedActivityView": {
        "additionalProperties": false,
        "properties": {
          "area_or_address": {
            "title": "Area Or Address",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "enum": [
              "可直接查看",
              "地点待确认"
            ],
            "title": "Note",
            "type": "string"
          },
          "time_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Hint"
          }
        },
        "required": [
          "name",
          "area_or_address",
          "note"
        ],
        "title": "SharedActivityView",
        "type": "object"
      },
      "SharedDayView": {
        "additionalProperties": false,
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/SharedActivityView"
            },
            "title": "Activities",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "activities"
        ],
        "title": "SharedDayView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "把攻略文字整理成逐日行程卡片、路线和少量可直接采纳的建议。",
    "title": "BreezeTravel 行程查",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v3/me/data-consents": {
      "get": {
        "operationId": "get_data_consents_api_v3_me_data_consents_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConsentView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Data Consents",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
    "/api/v3/me/data-consents/{purpose}": {
      "put": {
        "operationId": "set_data_consent_api_v3_me_data_consents__purpose__put",
        "parameters": [
          {
            "in": "path",
            "name": "purpose",
            "required": true,
            "schema": {
              "enum": [
                "memory",
                "feedback",
                "training-eval"
              ],
              "title": "Purpose",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConsentView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Data Consent",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
    "/api/v3/me/shares": {
      "get": {
        "operationId": "list_my_shares_api_v3_me_shares_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShareListItemView"
                  },
                  "title": "Response List My Shares Api V3 Me Shares Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List My Shares",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
    "/api/v3/me/shares/{share_ref}": {
      "delete": {
        "operationId": "revoke_my_share_api_v3_me_shares__share_ref__delete",
        "parameters": [
          {
            "in": "path",
            "name": "share_ref",
            "required": true,
            "schema": {
              "title": "Share Ref",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke My Share",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
...
    "/api/v3/me/travel-preferences": {
      "delete": {
        "operationId": "clear_travel_preferences_api_v3_me_travel_preferences_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Travel Preferences",
        "tags": [
          "memory-share-v3"
        ]
      },
      "get": {
        "operationId": "get_travel_preferences_api_v3_me_travel_preferences_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PreferenceMemoryView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Travel Preferences Api V3 Me Travel Preferences Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Travel Preferences",
        "tags": [
          "memory-share-v3"
        ]
      },
      "put": {
        "operationId": "save_travel_preferences_api_v3_me_travel_preferences_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferenceMemoryView"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferenceMemoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Travel Preferences",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
...
    "/api/v3/shares/{share_ref}": {
      "get": {
        "operationId": "read_shared_trip_api_v3_shares__share_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "share_ref",
            "required": true,
            "schema": {
              "title": "Share Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareProjectionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Shared Trip",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
    "/api/v3/shares/{share_ref}/exchange": {
      "post": {
        "operationId": "exchange_share_secret_api_v3_shares__share_ref__exchange_post",
        "parameters": [
          {
            "in": "path",
            "name": "share_ref",
            "required": true,
            "schema": {
              "title": "Share Ref",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange Share Secret",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
...
    "/api/v3/trip-understandings/{public_resource_id}/feedback": {
      "post": {
        "operationId": "record_trip_feedback_api_v3_trip_understandings__public_resource_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "public_resource_id",
            "required": true,
            "schema": {
              "title": "Public Resource Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackAcceptedView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Trip Feedback",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
...
    "/api/v3/trip-understandings/{public_resource_id}/shares": {
      "post": {
        "operationId": "create_trip_share_api_v3_trip_understandings__public_resource_id__shares_post",
        "parameters": [
          {
            "in": "path",
            "name": "public_resource_id",
            "required": true,
            "schema": {
              "title": "Public Resource Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareCreatedView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Trip Share",
        "tags": [
          "memory-share-v3"
        ]
      }
    },
...
  }
}